ref: muscle_controller
