+++ src/raspi/control/muscle/muscle_controller.py
@@ -1,9 +1,8 @@
 from dataclasses import dataclass, field
-from functools import wraps
 
 from raspi.io_controller import pressure_actions, valve_actions
 
-muscles = dict()
+from ...io_controller.util.device import create_device_global_store_definition
 
 
 @dataclass(slots=True)
@@ -36,121 +35,41 @@ class MuscleObj:
         setattr(self, key, value)
 
 
-def register_muscle(name: str, muscle) -> None:
-    """
-    Add a new muscle to the list of pins.
-
-    Args:
-        name (str): the name of the muscle
-        muscle (MuscleObj): the muscle dataclass object
-    """
-    if name in muscles:
-        raise ValueError(f"Muscle {name} already exists")
-    muscles[name] = muscle
-
-
-def get_muscle(
-    name: str,
-):
-    """
-    Get the pin number of a muscle.
-
-    Args:
-        name (str): the name of the muscle
-
-    Returns:
-        (MuscleObj) muscle dataclass object
-    """
-    return muscles[name]
-
-
-def get_muscle_names() -> list[str]:
-    """
-    Get a list of all the muscle names.
-
-    Returns:
-        a list of all the muscle names
-    """
-    return list(muscles.keys())
-
-
-def get_muscle_devices():
-    """
-    Get a list of all the muscle pins.
-
-    Returns:
-        a list of all the muscle pins
-    """
-    return list(muscles.values())
-
-
-def muscle_action(func: callable) -> callable:
-    """
-    Decorator for muscle actions.
-
-    Args:
-        func (callable): the function to decorate
-
-    Returns:
-        the decorated function
-    """
-
-    @wraps(func)
-    def wrapper(*args, **kwargs) -> bool:
-        """
-        Decorated function.
-
-        Args:
-            args: the arguments of the function
-            kwargs: the keyword arguments of the function
-        """
-        if len(args) == 0:
-            raise ValueError("Muscle name not provided")
-        if not isinstance(args[0], str) and not isinstance(args[0], MuscleObj):
-            raise ValueError("Muscle name must be a string or a MuscleObj")
-        muscle = get_muscle(args[0]) if isinstance(args[0], str) else args[0]
-        res = func(muscle, *args[1:], **kwargs)
-        if res is None:
-            return True
-        return res
-
-    return wrapper
+(muscle_action,) = create_device_global_store_definition([MuscleObj])("muscle")
+__all__ = ["contract", "relax", "toggle_muscle_state"]
 
 
 @muscle_action
-def contract(muscle) -> bool:
+def contract(muscle, check=pressure_actions.is_pressure_ok) -> bool:
     """
     Contract a muscle.
 
     Args:
         muscle (Muscle): the muscle to contract
+        check (callable): the function to check the pressure (default: is_pressure_ok from pressure_actions)
 
     Returns:
         True if the muscle was contracted, False otherwise
     """
-    pressure = pressure_actions.is_pressure_ok(muscle.pressure)
-    if pressure:
+    if not check(muscle.pressure):
         print(f"{muscle.pressure}: Pressure check failed, cannot contract muscle")
         return False
-    print("contracting muscle")
     valve_actions.turn_valve_on(muscle.valve)
     return True
 
 
 @muscle_action
-def relax(muscle) -> bool:
+def relax(muscle, check=valve_actions.get_valve_state) -> bool:
     """
     Relax a muscle.
 
     Args:
         muscle (Muscle): the muscle to relax
-        check (callable): the function to check the state of the valve (default: get_valve_state from io/pneumatics/valve.py)
+        check (callable): the function to check the state of the valve (default: get_valve_state from valve_actions)
     """
-    check = not valve_actions.get_valve_state(muscle.valve)
-    if check:
+    if not check(muscle.valve):
         print(f"{muscle.valve}: Valve check failed, Muscle is already relaxed")
         return False
-    print("relaxing muscle")
     valve_actions.turn_valve_off(muscle.valve)
     return True
 
@@ -167,11 +86,3 @@ def toggle_muscle_state(muscle) -> bool:
     if is_contracted:
         return relax(muscle)
     return contract(muscle)
-
-
-def noop(*args, **kwargs) -> bool:
-    """
-    No operation function.
-    Pass into check if you want to disable the check.
-    """
-    return True
